Refactor CommmitAvatar struct
