+++ crates/editor/src/blame_entry_tooltip.rs
@@ -1,0 +1,246 @@
+use futures::Future;
+use git::blame::BlameEntry;
+use git::Oid;
+use gpui::{
+    Asset, Element, ParentElement, Render, ScrollHandle, StatefulInteractiveElement, WeakView,
+    WindowContext,
+};
+use settings::Settings;
+use std::hash::Hash;
+use theme::{ActiveTheme, ThemeSettings};
+use ui::{
+    div, h_flex, tooltip_container, v_flex, Avatar, Button, ButtonStyle, Clickable as _, Color,
+    FluentBuilder, IconName, IconPosition, InteractiveElement as _, IntoElement, SharedString,
+    Styled as _, ViewContext,
+};
+use ui::{ButtonCommon, Disableable as _};
+use workspace::Workspace;
+
+use crate::git::blame::{CommitDetails, GitRemote};
+use crate::EditorStyle;
+
+struct CommitAvatar<'a> {
+    details: Option<&'a CommitDetails>,
+    sha: Oid,
+}
+
+impl<'a> CommitAvatar<'a> {
+    fn new(details: Option<&'a CommitDetails>, sha: Oid) -> Self {
+        Self { details, sha }
+    }
+}
+
+impl<'a> CommitAvatar<'a> {
+    fn render(&'a self, cx: &mut ViewContext<BlameEntryTooltip>) -> Option<impl IntoElement> {
+        let remote = self
+            .details
+            .and_then(|details| details.remote.as_ref())
+            .filter(|remote| remote.hosts_supports_avatars())?;
+
+        let avatar_url = CodeHostAvatarForSha::new(remote.clone(), self.sha);
+
+        cx.with_element_context(|cx| {
+            match cx.use_cached_asset::<CodeHostAvatarForSha>(&avatar_url) {
+                // Loading or no avatar found
+                None | Some(None) => Some(div().w(gpui::rems(1.)).into_any()),
+                // Found
+                Some(Some(url)) => Some(Avatar::new(url.to_string()).into_element().into_any()),
+            }
+        })
+    }
+}
+
+#[derive(Clone, Debug)]
+struct CodeHostAvatarForSha {
+    sha: Oid,
+    remote: GitRemote,
+}
+
+impl Hash for CodeHostAvatarForSha {
+    fn hash<H: std::hash::Hasher>(&self, state: &mut H) {
+        self.sha.hash(state);
+        self.remote.code_host.hash(state);
+    }
+}
+
+impl CodeHostAvatarForSha {
+    fn new(remote: GitRemote, sha: Oid) -> Self {
+        Self { remote, sha }
+    }
+}
+
+impl Asset for CodeHostAvatarForSha {
+    type Source = Self;
+    type Output = Option<SharedString>;
+
+    fn load(
+        source: Self::Source,
+        cx: &mut WindowContext,
+    ) -> impl Future<Output = Self::Output> + Send + 'static {
+        let client = cx.http_client();
+
+        async move {
+            source
+                .remote
+                .avatar_url(source.sha, client)
+                .await
+                .map(|url| SharedString::from(url.to_string()))
+        }
+    }
+}
+
+pub(crate) struct BlameEntryTooltip {
+    blame_entry: BlameEntry,
+    details: Option<CommitDetails>,
+    editor_style: EditorStyle,
+    workspace: Option<WeakView<Workspace>>,
+    scroll_handle: ScrollHandle,
+}
+
+impl BlameEntryTooltip {
+    pub(crate) fn new(
+        blame_entry: BlameEntry,
+        details: Option<CommitDetails>,
+        style: &EditorStyle,
+        workspace: Option<WeakView<Workspace>>,
+    ) -> Self {
+        Self {
+            editor_style: style.clone(),
+            blame_entry,
+            details,
+            workspace,
+            scroll_handle: ScrollHandle::new(),
+        }
+    }
+}
+
+impl Render for BlameEntryTooltip {
+    fn render(&mut self, cx: &mut ViewContext<Self>) -> impl IntoElement {
+        let avatar = CommitAvatar::new(self.details.as_ref(), self.blame_entry.sha).render(cx);
+
+        let author = self
+            .blame_entry
+            .author
+            .clone()
+            .unwrap_or("<no name>".to_string());
+
+        let author_email = self.blame_entry.author_mail.clone();
+
+        let pretty_commit_id = format!("{}", self.blame_entry.sha);
+        let short_commit_id = pretty_commit_id.chars().take(6).collect::<String>();
+        let absolute_timestamp = blame_entry_absolute_timestamp(&self.blame_entry, cx);
+
+        let message = self
+            .details
+            .as_ref()
+            .map(|details| {
+                crate::render_parsed_markdown(
+                    "blame-message",
+                    &details.parsed_message,
+                    &self.editor_style,
+                    self.workspace.clone(),
+                    cx,
+                )
+                .into_any()
+            })
+            .unwrap_or("<no commit message>".into_any());
+
+        let ui_font_size = ThemeSettings::get_global(cx).ui_font_size;
+        let message_max_height = cx.line_height() * 12 + (ui_font_size / 0.4);
+
+        tooltip_container(cx, move |this, cx| {
+            this.occlude()
+                .on_mouse_move(|_, cx| cx.stop_propagation())
+                .child(
+                    v_flex()
+                        .w(gpui::rems(30.))
+                        .gap_4()
+                        .child(
+                            h_flex()
+                                .gap_x_2()
+                                .overflow_x_hidden()
+                                .flex_wrap()
+                                .children(avatar)
+                                .child(author)
+                                .when_some(author_email, |this, author_email| {
+                                    this.child(
+                                        div()
+                                            .text_color(cx.theme().colors().text_muted)
+                                            .child(author_email),
+                                    )
+                                })
+                                .border_b_1()
+                                .border_color(cx.theme().colors().border),
+                        )
+                        .child(
+                            div()
+                                .id("inline-blame-commit-message")
+                                .occlude()
+                                .child(message)
+                                .max_h(message_max_height)
+                                .overflow_y_scroll()
+                                .track_scroll(&self.scroll_handle),
+                        )
+                        .child(
+                            h_flex()
+                                .text_color(cx.theme().colors().text_muted)
+                                .w_full()
+                                .justify_between()
+                                .child(absolute_timestamp)
+                                .child(
+                                    Button::new("commit-sha-button", short_commit_id.clone())
+                                        .style(ButtonStyle::Transparent)
+                                        .color(Color::Muted)
+                                        .icon(IconName::FileGit)
+                                        .icon_color(Color::Muted)
+                                        .icon_position(IconPosition::Start)
+                                        .disabled(
+                                            self.details.as_ref().map_or(true, |details| {
+                                                details.permalink.is_none()
+                                            }),
+                                        )
+                                        .when_some(
+                                            self.details
+                                                .as_ref()
+                                                .and_then(|details| details.permalink.clone()),
+                                            |this, url| {
+                                                this.on_click(move |_, cx| {
+                                                    cx.stop_propagation();
+                                                    cx.open_url(url.as_str())
+                                                })
+                                            },
+                                        ),
+                                ),
+                        ),
+                )
+        })
+    }
+}
+
+fn blame_entry_timestamp(
+    blame_entry: &BlameEntry,
+    format: time_format::TimestampFormat,
+    cx: &WindowContext,
+) -> String {
+    match blame_entry.author_offset_date_time() {
+        Ok(timestamp) => time_format::format_localized_timestamp(
+            timestamp,
+            time::OffsetDateTime::now_utc(),
+            cx.local_timezone(),
+            format,
+        ),
+        Err(_) => "Error parsing date".to_string(),
+    }
+}
+
+pub fn blame_entry_relative_timestamp(blame_entry: &BlameEntry, cx: &WindowContext) -> String {
+    blame_entry_timestamp(blame_entry, time_format::TimestampFormat::Relative, cx)
+}
+
+fn blame_entry_absolute_timestamp(blame_entry: &BlameEntry, cx: &WindowContext) -> String {
+    blame_entry_timestamp(
+        blame_entry,
+        time_format::TimestampFormat::MediumAbsolute,
+        cx,
+    )
+}
