+++ crates/git/Cargo.toml
@@ -23,6 +23,7 @@ sum_tree.workspace = true
 text.workspace = true
 time.workspace = true
 url.workspace = true
+util.workspace = true
 serde.workspace = true
 rope.workspace = true
 util.workspace = true
